lamdba/requirements.txt
botocore<1.13.0,>=1.5.0
boto3<1.10.0,>=1.4.4
#https://stackoverflow.com/questions/58243179/
#how-can-i-connect-to-snowflake-in-aws-lambda-function-without-using-snowflake-co
#snowflake-connector-python==2.0.1
snowflake-connector-python==1.9.1
asn1crypto<1.0.0
aws-sam-cli
awscli
#awscli==1.16.263
# ERROR: aws-sam-cli 0.23.0 has requirement six~=1.11.0, but you'll have six 1.12.0 which is incompatible.
six==1.11.0